fix(ms): add prefix ms in user-agent and move to catalog src
