+++ micro-services/src/ms-user-agent/src/data/configs/names.config.ts
