+++ micro-services/src/ms-user-agent/src/domain/contracts/service.contract.ts
