+++ micro-services/src/ms-user-agent/src/domain/models/browser.model.ts
