+++ micro-services/src/ms-user-agent/src/domain/models/device.model.ts
