+++ micro-services/src/ms-user-agent/src/domain/models/platform.model.ts
