+++ micro-services/src/ms-user-agent/src/domain/services/user-agent.service.ts
@@ -1,14 +1,14 @@
 import UserAgentRepository from "../../data/repositories/user-agent.repository"
 import type { Maybe } from "@windows98/toolkit/src/types"
 import type UserAgentServiceContract from "../contracts/service.contract"
-import type { TUserAgentDevices } from "../../domain/models/device.model"
+import type { TUserAgentDevices } from "../models/device.model"
 import type { TUserAgentBrowsers } from "../models/browser.model"
 import type { TUserAgentAllPlatforms } from "../models/platform.model"
 
 export default class UserAgentService implements UserAgentServiceContract {
 	constructor(
-    private readonly userAgentRepository = new UserAgentRepository(),
-	) {}
+		private readonly userAgentRepository = new UserAgentRepository(),
+	) { }
 
 	public getBrowserName(): Maybe<TUserAgentBrowsers> {
 		return this.userAgentRepository.getBrowserName()
