+++ micro-services/src/ms-user-agent/tests/mocks/user-agent.mock.ts
